Rename files and move media type extractor into buildSettings
