+++ src/app/components/MediaLoader/configs/aresMedia.ts
