+++ src/app/components/MediaLoader/configs/clipMedia.ts
