+++ src/app/components/MediaLoader/index.tsx
@@ -13,7 +13,6 @@ import Placeholder from './Placeholder';
 import getProducerFromServiceName from './utils/getProducerFromServiceName';
 import getCaptionBlock from './utils/getCaptionBlock';
 import styles from './index.styles';
-import getMediaBlockType from './utils/getMediaBlockType';
 
 const logger = nodeLogger(__filename);
 
@@ -74,12 +73,7 @@ const MediaLoader = ({ blocks, className }: Props) => {
 
   const producer = getProducerFromServiceName(service);
 
-  const mediaBlockType = getMediaBlockType(blocks);
-
-  if (!mediaBlockType) return null;
-
   const config = buildConfig({
-    mediaBlockType,
     blocks,
     counterName,
     statsDestination,
